requirements.txt
jax[cpu]
# Used for image debug utilities
numpy
Pillow
# Used for visualization
matplotlib